+++ RELEASES.md
@@ -1,5 +1,7 @@
-Version tw-0.8.9-3.9 (2022-02-01)
+Version tw-0.8.9-3.9 (2022-02-04)
 ========================
+ * NEW: Tap driver monitoring icon to toggle screen dimming
+   * Screen will automatically undim for warning/critical alerts
  * NEW: [Volt] [optional] control coasting for cruise and one-pedal mode using D and L mode with the gear shifter
    * Coasts in D, regen in L
    * Two separate toggles
@@ -7,6 +9,7 @@ Version tw-0.8.9-3.9 (2022-02-01)
  * NEW: [Volt] [optional] tie one-pedal engage on gas setting to D/L position: off in D; on in L
    * Suggested by Shadowlight5
  * NEW: Display vehicle fingerprint on startup
+ * NEW: UI metrics; Volt high voltage battery voltage, wattage, amperage, and volt+watt
  * IMPROVED: Brake indicator now shows regen/engine braking while cruising, or regen while in one-pedal regen mode
  * IMPROVED: UI Metrics now larger when four or fewer present (suggested by C Tyrell)
  * FIXED: [Volt] Prevent misprinting as Escalade ESV